+++ packages/orq-rc/src/models/operations/createrouterresponse.ts
@@ -565,10 +565,10 @@ export type CreateRouterResponseResponseBody = {
   completedAt: number | null;
   conversation?: components.ConversationParam | undefined;
   createdAt: number;
-  error: components.ResponseError;
+  error: components.ResponseError | null;
   frequencyPenalty: number;
   id: string;
-  incompleteDetails: components.IncompleteDetails;
+  incompleteDetails: components.IncompleteDetails | null;
   /**
    * Array of input items (messages, function call outputs, etc.)
    */
@@ -595,7 +595,7 @@ export type CreateRouterResponseResponseBody = {
   previousResponseId: string | null;
   promptCacheKey: string | null;
   promptCacheRetention: string | null;
-  reasoning: components.Reasoning;
+  reasoning: components.Reasoning | null;
   safetyIdentifier: string | null;
   serviceTier: ServiceTier;
   status: CreateRouterResponseStatus;
@@ -1413,10 +1413,10 @@ export const CreateRouterResponseResponseBody$inboundSchema: z.ZodType<
   completed_at: z.nullable(z.number().int()),
   conversation: components.ConversationParam$inboundSchema.optional(),
   created_at: z.number().int(),
-  error: components.ResponseError$inboundSchema,
+  error: z.nullable(components.ResponseError$inboundSchema),
   frequency_penalty: z.number(),
   id: z.string(),
-  incomplete_details: components.IncompleteDetails$inboundSchema,
+  incomplete_details: z.nullable(components.IncompleteDetails$inboundSchema),
   input: z.nullable(z.array(z.any())),
   instructions: z.nullable(z.string()),
   max_output_tokens: z.nullable(z.number().int()),
@@ -1431,7 +1431,7 @@ export const CreateRouterResponseResponseBody$inboundSchema: z.ZodType<
   previous_response_id: z.nullable(z.string()),
   prompt_cache_key: z.nullable(z.string()),
   prompt_cache_retention: z.nullable(z.string()),
-  reasoning: components.Reasoning$inboundSchema,
+  reasoning: z.nullable(components.Reasoning$inboundSchema),
   safety_identifier: z.nullable(z.string()),
   service_tier: ServiceTier$inboundSchema,
   status: CreateRouterResponseStatus$inboundSchema,