+++ packages/orq-rc/RELEASES.md
@@ -9099,3 +9099,13 @@ Based on:
 - [typescript v4.9.0-rc.49] packages/orq-rc
 ### Releases
 - [NPM v4.9.0-rc.49] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.49 - packages/orq-rc
+
+## 2026-05-12 12:36:42
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.9.0-rc.50] packages/orq-rc
+### Releases
+- [NPM v4.9.0-rc.50] https://www.npmjs.com/package/@orq-ai/node/v/4.9.0-rc.50 - packages/orq-rc